Add linting

+++ contracts/ErrorReporter.sol
@@ -50,27 +50,31 @@ contract ComptrollerErrorReporter {
     }
 
     /**
-      * @dev `error` corresponds to enum Error; `info` corresponds to enum FailureInfo, and `detail` is an arbitrary
-      * contract-specific code that enables us to report opaque error codes from upgradeable contracts.
-      **/
-    event Failure(uint error, uint info, uint detail);
+     * @dev `error` corresponds to enum Error; `info` corresponds to enum FailureInfo, and `detail` is an arbitrary
+     * contract-specific code that enables us to report opaque error codes from upgradeable contracts.
+     **/
+    event Failure(uint256 error, uint256 info, uint256 detail);
 
     /**
-      * @dev use this when reporting a known error from the money market or a non-upgradeable collaborator
-      */
-    function fail(Error err, FailureInfo info) internal returns (uint) {
-        emit Failure(uint(err), uint(info), 0);
+     * @dev use this when reporting a known error from the money market or a non-upgradeable collaborator
+     */
+    function fail(Error err, FailureInfo info) internal returns (uint256) {
+        emit Failure(uint256(err), uint256(info), 0);
 
-        return uint(err);
+        return uint256(err);
     }
 
     /**
-      * @dev use this when reporting an opaque error from an upgradeable collaborator contract
-      */
-    function failOpaque(Error err, FailureInfo info, uint opaqueError) internal returns (uint) {
-        emit Failure(uint(err), uint(info), opaqueError);
+     * @dev use this when reporting an opaque error from an upgradeable collaborator contract
+     */
+    function failOpaque(
+        Error err,
+        FailureInfo info,
+        uint256 opaqueError
+    ) internal returns (uint256) {
+        emit Failure(uint256(err), uint256(info), opaqueError);
 
-        return uint(err);
+        return uint256(err);
     }
 
     error InvalidCollateralFactor();
@@ -87,13 +91,22 @@ contract ComptrollerErrorReporter {
      *   a predefined threshold. In this case only batch liquidations (either liquidateAccount
      *   or healAccount) are available.
      */
-    error MinimalCollateralViolated(uint256 expectedGreaterThan, uint256 actual);
-    error CollateralExceedsThreshold(uint256 expectedLessThanOrEqualTo, uint256 actual);
-    error InsufficientCollateral(uint256 collateralToSeize, uint256 availableCollateral);
+    error MinimalCollateralViolated(
+        uint256 expectedGreaterThan,
+        uint256 actual
+    );
+    error CollateralExceedsThreshold(
+        uint256 expectedLessThanOrEqualTo,
+        uint256 actual
+    );
+    error InsufficientCollateral(
+        uint256 collateralToSeize,
+        uint256 availableCollateral
+    );
 }
 
 contract TokenErrorReporter {
-    uint public constant NO_ERROR = 0; // support legacy return codes
+    uint256 public constant NO_ERROR = 0; // support legacy return codes
 
     error TransferComptrollerRejection(uint256 errorCode);
     error TransferNotAllowed();
