+++ contracts/test/TetherInterface.sol
@@ -4,5 +4,5 @@ pragma solidity ^0.8.10;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface TetherInterface is IERC20 {
-    function setParams(uint newBasisPoints, uint newMaxFee) external;
+    function setParams(uint256 newBasisPoints, uint256 newMaxFee) external;
 }
